Added s3 key in frontend
- 2 input file options (raw/s3) in create new dry run
